refactored order crossover tests
